make the filter cut off a double so it doesn't narrow

+++ pedals/filter_pedal.h
@@ -16,18 +16,18 @@ class FilterPedal : public Pedal {
     PedalInfo info;
     info.name = GetName();
 
-    info.knobs = {PedalKnob{.name = "cut_off_hz",
-                            .value = static_cast<double>(cut_off_hz_),
-                            .tweak_amount = 100},
-                  PedalKnob{.name = "sample_rate",
-                            .value = static_cast<double>(sample_rate_),
-                            .tweak_amount = 100}};
+    info.knobs = {
+        PedalKnob{
+            .name = "cut_off_hz", .value = cut_off_hz_, .tweak_amount = 100},
+        PedalKnob{.name = "sample_rate",
+                  .value = static_cast<double>(sample_rate_),
+                  .tweak_amount = 100}};
     return info;
   }
 
   void AdjustKnob(const PedalKnob& pedal_knob) override {
     if (pedal_knob.name == "cut_off_hz") {
-      cut_off_hz_ = static_cast<int>(pedal_knob.value);
+      cut_off_hz_ = pedal_knob.value;
     } else if (pedal_knob.name == "sample_rate") {
       sample_rate_ = static_cast<unsigned int>(pedal_knob.value);
     }
@@ -37,7 +37,7 @@ class FilterPedal : public Pedal {
 
  private:
   unsigned int sample_rate_ = 44100;
-  int cut_off_hz_ = 1000;
+  double cut_off_hz_ = 1000;
   FilterType filter_{cut_off_hz_, sample_rate_};
 };
